fix: move createClient into useEffect
move createClient into a useEffect to keep everything at client side when building the project
