+++ frontend/src/context/AuthContext.tsx
@@ -1,7 +1,6 @@
 'use client';
-
-import { createContext, useContext, useEffect, useState, ReactNode } from 'react';
-import { createClient } from '../lib/supabase/client';
+import { createContext, useContext, ReactNode } from 'react';
+import { AuthProviderInner } from './AuthProviderInner';
 
 type AuthContextType = {
   user: { name: string; email: string } | null;
@@ -11,54 +10,14 @@ type AuthContextType = {
 
 const AuthContext = createContext<AuthContextType | undefined>(undefined);
 
-export function AuthProvider({ children }: { children: ReactNode }) {
-  const [user, setUser] = useState<{ name: string; email: string } | null>(null);
-  const [isLoading, setIsLoading] = useState(true);
-  const supabase = createClient();
-
-  useEffect(() => {
-    const initAuth = async () => {
-      const {
-        data: { session },
-      } = await supabase.auth.getSession();
-      if (session?.user) {
-        setUser({
-          name: session.user.user_metadata.full_name ?? session.user.email ?? 'User',
-          email: session.user.email ?? '',
-        });
-      }
-      setIsLoading(false);
-    };
-
-    initAuth();
-
-    const { data: listener } = supabase.auth.onAuthStateChange((_event, session) => {
-      if (session?.user) {
-        setUser({
-          name: session.user.user_metadata.full_name ?? session.user.email ?? 'User',
-          email: session.user.email ?? '',
-        });
-      } else {
-        setUser(null);
-      }
-    });
-
-    return () => listener.subscription.unsubscribe();
-  }, [supabase]);
-
-  const logout = async () => {
-    setIsLoading(true);
-    await supabase.auth.signOut();
-    setUser(null);
-  };
-
-  return (
-    <AuthContext.Provider value={{ user, isLoading, logout }}>{children}</AuthContext.Provider>
-  );
-}
+export const AuthProvider = ({ children }: { children: ReactNode }) => (
+  <AuthProviderInner>{children}</AuthProviderInner>
+);
 
 export function useAuth() {
   const context = useContext(AuthContext);
   if (!context) throw new Error('useAuth must be used inside AuthProvider');
   return context;
 }
+
+export { AuthContext };
